Drone health, shooting, and shooting object logic
Drone now has health and will explode. Drone shoots at the player (Not damage the shield), and the object they shoot is slow but can be shot down. Adding a droneshot prefeb, alot of scripts under "testing" file, and changed the logic of the player cannonball to notice the drone and its shot. Still need to add it damaging the player and then work on the spawning logic.