Update projectile tags and pool manager objects
Fixed the Ally Power Up to work now: Removed the respawn timer logic so you have to force it to spawn itself and it will active a Empty that will host the ally turrets and start shooting at enemies to help the player. Duped the old powerup Mat so its green.

Updated the Ally turret perfab and script to now play sound when shooting and updated the logic so it is more random when shooting to make it more unique, to not have both turrets shooting at the same time.

Fixed some Bugs: Changed the CannonBall prefab by adding a 'PlayerShot' tag to it. Then changing some logic to check for this tag (like in the powerups), instead of checking the name of the object (so its now easier to switch out the cannon with a new model without breaking much).
And Expanded the PoolManager prefab inside the main game sense to include the projectiles the allies shoot.